+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
  - [NZXT E] Fix assertion of valid responses in retry loops
  - [HUE 2] Fix LED blinking during `status`
  - [HUE 2] Add missing identifier for 250 mm HUE 2 LED strips
+ - Restore experimental tag for the NZXT Kraken X31/X41/X61 family
 
 ## [1.3.0rc1] – 2019-11-03
 _Man page, Corsair RXi/HXi and NZXT E power supplies, Smart Device V2 and improved device discovery_